// cli-wallet/src/types.ts
export interface TransactionBuildParams {
  feePayer: Uint8Array; // 32-byte public key
  recentBlockhash: string;
  instructions: InstructionInput[];
}

export interface InstructionInput {
  programId: Uint8Array; // 32-byte public key
  keys: InstructionAccount[];
  data: Buffer;
}

export interface InstructionAccount {
  pubkey: Uint8Array; // 32-byte public key
  isSigner: boolean;
  isWritable: boolean;
}